+++ R/keypress.R
@@ -1,8 +1,8 @@
 
-#' Wait for a single keypress at the terminal
+#' Read a single keypress at the terminal
 #'
 #' It currently only works at Linux/Unix and OSX terminals,
-#' see \code{\link{has_keypress_support}}.
+#' and at the Windows command line. see \code{\link{has_keypress_support}}.
 #'
 #' The following special keys are supported:
 #' \itemize{
@@ -11,13 +11,22 @@
 #'   \item Function keys: from \sQuote{f1} to \sQuote{f12}.
 #'   \item Others: sQuote{home}, \sQuote{end},
 #'     \sQuote{insert}, \sQuote{delete}, \sQuote{pageup},
-#'     \sQuote{pagedown}.
+#'     \sQuote{pagedown}, \sQuote{tab}, \sQuote{enter},
+#'     \sQuote{backspace} (same as \sQuote{delete} on OSX keyboards),
+#'     \sQuote{escape}.
+#'   \item Control with one of the following keys: \sQuote{a}, \sQuote{b},
+#'     \sQuote{c}, \sQuote{d}, \sQuote{e}, \sQuote{f}, \sQuote{h},
+#'     \sQuote{k}, \sQuote{l}, \sQuote{n}, \sQuote{p}, \sQuote{t},
+#'     \sQuote{u}, \sQuote{w}.
 #' }
 #'
-#' @return The key pressed, a character scalar.
+#' @param block Whether to wait for a key press, if there is none
+#'   available now.
+#' @return The key pressed, a character scalar. For non-blocking reads
+#'   NA is returned if no keys are available.
 #'
 #' @family keypress
-#' @useDynLib keypress
+#' @useDynLib keypress, .registration = TRUE, .fixes = "C_"
 #' @export
 #' @examples
 #' \dontrun{
@@ -25,11 +34,13 @@
 #' cat("You pressed key", x, "\n")
 #' }
 
-keypress <- function() {
+keypress <- function(block = TRUE) {
   if (!has_keypress_support()) {
     stop("Your platform/terminal does not support keypress")
   }
-  .Call("keypress", PACKAGE = "keypress")
+  block <- as.logical(block)
+  if (length(block) != 1) stop("'block' must be a logical scalar")
+  .Call(C_keypress, block)
 }
 
 #' Check if the current platform/terminal supports reading