+++ inst/NEWS.md
@@ -1,4 +1,16 @@
 
+# 1.1.1
+
+- Use native routine registration to avoid R CMD check warnings.
+- Make sure compilation units are not empty, to avoid another R CMD check
+  warning.
+
+# 1.1.0
+
+- Non-blocking reads.
+- Better reading individual keys from a copy-paste, or the OS buffer
+  in general.
+
 # 1.0.0
 
 First released version.